Make the manifest reader into a cargo so we can re-use them. Rename some main rust files to core.rs
