+++ indexer/json-rpc-server/Cargo.toml
@@ -8,6 +8,6 @@ edition = "2018"
 [dependencies]
 jsonrpc-http-server = "15.0"
 serde = { version = "1.0", features = ["derive"] }
-serde_yaml = "0.8"
 serde_json = "1.0.40"
-anyhow = "1.0"
+# Massbit dependencies
+manifest-reader = { path = "../manifest-reader" }
