+++ indexer/manifest-reader/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "manifest-reader"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.24"
+anyhow = "1.0"
+serde_yaml = "0.8"
