+++ indexer/manifest-reader/src/lib.rs
